+++ src/App.tsx
@@ -14,6 +14,7 @@ function App() {
 				<a
 					href="https://vite.dev"
 					target="_blank"
+					rel="noreferrer"
 				>
 					<img
 						src={viteLogo}
@@ -24,6 +25,7 @@ function App() {
 				<a
 					href="https://react.dev"
 					target="_blank"
+					rel="noreferrer"
 				>
 					<img
 						src={reactLogo}
@@ -34,7 +36,13 @@ function App() {
 			</div>
 			<h1>Vite + React</h1>
 			<div className="card">
-				<button onClick={() => setCount((count) => count + 1)}>count is {count}</button>
+				<button
+					onClick={() => {
+						setCount((count) => count + 1);
+					}}
+				>
+					count is {count}
+				</button>
 				<p>
 					Edit <code>src/App.tsx</code> and save to test HMR
 				</p>